added Calculator.java and RecordsHandler.java also changed some data types

+++ src/main/java/AgeFactorRec.java
@@ -1,2 +1,2 @@
-public record AgeFactorRec(int max, float factor) {
+public record AgeFactorRec(int max, double factor) {
 }
